In tests, use [% and %] as delimiters for Text::Template.
The standard delimiter setting in Text::Template is curly braces. They
have several disadvantages. Changes the delimiters in all test
specifications.

The new setting is the default for Template::Toolkit, which we may use
in the future. For now, we continue to use Text::Template.

The biggest disadvantage of the curly braces is that variables in
Debian control files, such as ${misc:Depends}, must be escaped.  The
resulting backslashes are cumbersome, visually unattractive, and error
prone.

Another disadvantage may be speed, although probably not paramount for
our application. The module's author wrote:

    "Because the parsing of templates is simplified by the absence of
    backslash escapes, using alternative DELIMITERS may speed up the
    parsing process by 20-25%. This shows that my original choice of
    { and } was very bad."

This replaces the delimiters for template substitutions in Lintian's
test specifications.

The reporting framework also uses Text::Template but is not affected.