Provide a new role called Lintian::Tag::Issuer; facilities for issuing tags.
This a role for modules that issue tags. The tags are collected for
later processing, which allows them to be sorted before they are
issued.

The functionality is ultimately needed to make Lintian's output
independent from the order in which checks run. The execution order is
a poor and unstable odering method.

Gbp-Dch: ignore