Properly calculate an appropriate prior changelog version for source NMUs. (Closes: #942411)
This currently only works for the first source NMU. For subsequent
non-maintainer uploads, the check is effectively disabled.

The best way to solve it for more cases is probably to do some version
math when the critical parts are numerical (\d+), and then reassemble
the expected version from the parts.