Rename Lintian::Processable::Group, use Moo; add many clean ups and refactorings.
Much of the functionality regarding groups was consolidated in this
class. A major source was commands/lintian.pm, which is now much
shorter.

Rationalizes the addition of new "processables" to groups, for which
there were previously three different algorithms in various parts of
the code base.  The best way to add a file to the pool is to add it to
a newly created group, and add that to the pool.

An important realization was that the group, by virtue of its ability
to draw connections between source and installation packages, is the
primary unit being checked by Lintian.

Renames the class to Lintian::Processable::Group from the previous
Lintian::ProcessableGroup (without colons).

Converts the Group class to use the Moo object system.

The changes are incomplete and will continue as global variables are
being factored out.

Gbp-Dch: ignore