Rename Lintian::Processable::Pool, use Moo; add many clean ups and refactorings.
Much of the functionality regarding pools was consolidated in this
class. A major source was commands/lintian.pm. The goal is hide most
processing from higher code layers, which will make it less error
prone and more readable.

The Lintian::Processable::Group elements contained in this pool are
the primary unit Lintian runs on.  The best way to add a file to the
pool is to add it to a newly created group, and then add it to the
pool.

Renames the class to Lintian::Processable::Pool from the previous
Lintian::ProcessablePool (without colons).

Converts the Pool class to use the Moo object system.

The changes are incomplete and will continue as global variables are
being factored out.

Gbp-Dch: ignore