In commands/lintian, use new class Lintian::Processable::Pool.
Much of the execution moved to library modules. The result is a more
readable file in Lintian's master script.

Procedural details are now hidden in appropriate layers.

The objective is to make the Lintian code base easier to read for new
contributors.

The changes are incomplete and will continue as global variables are
being factored out.

Gbp-Dch: ignore