Add new tag 'manpage-in-udeb'.
This tag replaces 'udeb-contains-documentation-file' when pertaining
to manual pages.

A separate tag was necessary because the same tag was issued by two
different parts that were split off from the old check 'files'.
Lintian currently does not allow two checks to issue the same tag.

The tag appears to be untested in the test suite.

Gbp-Dch: ignore