In private/generate-profiles.pl, extract auto-reject tags from data posted by FTP Master.
The script downloads the most recent data for FTP Master auto
rejects from their URL at:

    https://ftp-master.debian.org/static/lintian.tags

The resulting profile is not used by the FTP Master team, but is made
available to users via the command line option -F. For details, please
see Bug#943536.

The script previously used local copies of similar data but it was
outdated and was in an unfamiliar format. The old, static copies of
that data will be removed from the Git repo in the near future.

The module LWP::Simple, which is used to download, was not added as a
prerequisite to d/control. The script is intended to be run manually
by the development team, and is not shipped to users.

The data file from FTP Master is in YAML format, and should be parsed
as such. The current approach via regular expressions is not ideal.