Update Master FTP auto-reject profile with latest data.
This incorporates the most recent information from the offical
repository for the data at:

    https://ftp-master.debian.org/static/lintian.tags

This profile is not used by the FTP Master team, but is made available
to users via the command line option -F, presumably to check planned
uploads for potential rejection. For details, please see Bug#943536.

The profile was outdated and has been updated.

Gbp-Dch: ignore