In Test::Lintian::Prepare, record runpath in desc working copy.
When preparing test sources, records the path to the source directory
among the test data. This path is later used in Makefiles to access
sources while building in another folder. It makes it possible to
build in a directory other than the source directory.

This commit is part of a series to build test packages adaptively and
as needed, instead of in bulk. The approach will involve a separation
of test specifications into building and running Lintian. A rebuild is
required only when the build spec has changed (but the form of Lintian
invocation does not matter).

Gbp-Dch: ignore