In skeleton upload-non-native, automatically fill pre-upstream.in with values, if present.
An upcoming adjustment to the test cruft/cruft-empty-diff will turn a
file pre-upstream into a template in order to have the path for the
package sources available. Makes the templating engine aware of the
need to fill that file with values.

This commit is part of a series to build test packages adaptively and
as needed, instead of in bulk. The approach will involve a separation
of test specifications into building and running Lintian. A rebuild is
required only when the build spec has changed (but the form of Lintian
invocation does not matter).

Gbp-Dch: ignore