In tests, rename pre_build to pre-build.
For historical reasons, some of the test helper scripts such as
pre_build had underscores in their names while others had hyphens.
This commit is part of a series to standardize the names using
hyphens.

Hyphens can be easier to read but also have drawbacks. Significant in
this context is that Makefile variables may not contain hyphens.

Gbp-Dch: ignore