Fix template whitelists for the migration from underscores to hyphens in test helper scripts.
For historical reasons, some of the test helper scripts such as
pre_build had underscores in their names while others had hyphens.
This commit is part of a series to standardize the names using
hyphens.

This adjusts the template whitelists to allow filling of the renamed
files.

Hyphens can be easier to read but also have drawbacks. Significant in
this context is that Makefile variables may not contain hyphens.

Gbp-Dch: ignore