In test suite, adjust builder templates to different paths for sources and destinations.
Makes most paths in the make-based package builders for the test suite
relative to either the build sources or the build destination. The
modifications are similar to those made when adapting Makefiles to
VPATH, although that feature is not used here currently.

This commit is part of a series to build test packages adaptively and
as needed, instead of in bulk. The approach will involve a separation
of test specifications into building and running Lintian. A rebuild is
required only when the build spec has changed (but the form of Lintian
invocation does not matter).

Gbp-Dch: ignore