Adjust test debian/changelog/changelog-file-unreleased-signed-changes for adaptive building.
Make the modifications necessary to get this test to work with the new
templates, which allow package builds and packages source to reside in
different locations.

This commit is part of a series to build test packages adaptively and
as needed, instead of in bulk. The approach will involve a separation
of test specifications into building and running Lintian. A rebuild is
required only when the build spec has changed (but the form of Lintian
invocation does not matter).

Gbp-Dch: ignore