Revert "Emit a warning for overly-long binary package fields. (Closes: #942493)"
This reverts commit dab49016e0caa5a28faa3d4ffa7d0b7862095586.

There were two issues with this commit:

    (1) The edit to the test included a duplicate long description. It
triggered a spurious tag that should not have found its way into the
test suite.  The tag can be seen from the diff on 'tags' in the
original commit. (The extra description is harder to see because the
other fields are so long, but it is there.)

    (2) In the future, please do not modify existing tests unless you
modify existing tags. It is easy to create new test cases: Just copy a
template to modify from t/templates/debian-native (or nearby) and
modify an existing 'desc' file from a test nearby. Here, two new tests
can be added.

Keeping tests specific helps us when moving tags around to make
Lintian better for everyone. The check will be re-added with those
minor improvements in the near future.

Thank you to Sylvestre Ledru <sylvestre@debian.org> for your
valuable contributions to Lintian!