In collection strings, inline coll/strings-helper to avoid spinning up another Perl.
In a surprising development, this effort resulted in the elimination
of the helper code. For reasons unknown, the helper did exactly what
was applied manually to some selected 'interesting' binaries here.

The only difference was that list produced by the helper eliminated
duplicates. 'strings' from GNU binutils does not do that.  After some
consideration, the 'unique' feature was dropped.

The helper also called 'gzip' in parallel, but it is not clear that it
produced any speed advantage. The disadvantage, on the other hand, was
that the script started a new Perl interpreter for each installation
package.

This avoids starting one Perl interpreter per processable (source or
installation packages). It should save a few minutes in Gitlab.

Gbp-Dch: ignore