In collection/md5sums, store values in BerkeleyDB.
The database will help with irregular file names. In order to further
reduce issues with irregular file names, this now also uses the --zero
option to md5sum.

An earlier attempt to use it (commit 59acd2c2) had to be reverted.
The --zero option did not appear until version 8.30 of the GNU
coreutils, which were not available in Debian stable at the time.

The reference to coreutils may be important enough for the changelog.
The version requirement was already added to d/control, but only in
the Depends fields (and not in Build-Depends).