Clarify variable names in check 'md5sums'; distinguish between supplied and calculated information.
The naming of variables in this check is difficult because so many
things are called 'md5sums'. This commit tries to distiguish more
clearly between information supplied by the package being checked and
information calculated by Lintian.

This would have fixed Bug#943745, except that the bug existed only in
the author's mind.

Gbp-Dch: ignore