Add a minimum version requirement of 8.30 for coreutils to d/control.
In the near future, the collection script md5sums will use the --zero
option to generate output. That command line option is not available
in coreutils versions prior to 8.30.

Adds the version requirement to d/control, but only in Depends and not
in Build-Depends.

An earlier attempt to use the --zero option (commit 59acd2c2) had to
be reverted because it was not available in Debian stable at the time.

This is part of a larger effort to accomodate irregular input file
names without escape sequences. Was prompted by 929729, but does not
yet solve it.