In collection src-orig-index, drop reference to obsolete class Processable::Package.
Together with Lab::Entry, this class was merged into Processable. All
three did basically the same thing. The differences will be worked out
with Moo roles in the future.

The presence of this reference in a collection script caused a test
failure on Gitlab after all tests completed successfully on a desktop
computer running Debian stable.

It probably means that the work product of the collection script
src-orig-index is not tested.

Together with the following commit, this should restore building on
Gitlab.

Gbp-Dch: ignore