Require libclass-xsaccessor-perl in both Depends and Build-Depends.
Moo accessors behave somewhat differently when Class::XSAccessor is
available. Bug#943724 only showed up when the module was installed.
This commit requires the module for everyone.

Moo's behavior is documented here:

    https://metacpan.org/pod/Moo#MOO-AND-CLASS::XSACCESSOR

While it is not great for Lintian to depend on modules that are not
essential, doing so will reduce the risk for malfunction or FTBFS in
this case. Many people run builders with recommended packages enabled.

On the positive side, Class::XSAccessor should be 3-4 times faster,
although it may not be relevant for Lintian. The POD tests are also
more stringent, which is probably desirable.

Lintian maintainers running Lintian primarily via t/bin/runtests or
frontend/lintian may not become aware of the new prerequisite without
a note.

A ticket was filed under RT#8012 to also update the installed packages
on lindsay.d.o.