In tests, set default to --pedantic.
This switches the default invocation of Lintian in the test suite to
'--pedantic'. Many new tags will appear, but only in related checks.

This commit was motivated by 4d88fce1. The test suite should not need
to be recalibrated because a tag's severity level changed. The whole
purpose of the 'universal' tag format was to remove this information.
This commit completes that train of thought.

Logically, this change also follows commit 0af8a3d8. In it, the test
suite ran just specified checks. It stopped many nuisance tags, such
as debian-watch-does-not-check-gpg-signature, from appearing
throughout the test suite. Now that tag appears only in tests related
to the check debian/watch.

The tests now focus on whether a condition was detected properly. It
is the essence of good check writing.

Some test specifications individually require the pedantic setting.
Those individual settings will be removed once it is clear that the
test suite performs as intended with this change.

Some tags relate to these default settings. They should be modified so
that the pedantic tags, like regular ones, appear less frequently and
ideally only when provoked by a particular test case.

Gbp-Dch: ignore