Convert Processable to Moo.
These changes look insignificant but they were a major step in
disentangling Processable from Collect and Unpacker.

The key was to wait with transferring the 'extra-fields' until the
info structure is read for the first time. At that point, the Unpacker
has completed all the unpacking.

All of this will fade into the distance as functionality can now be
moved between the appropriate places in various classes and roles.

Gbp-Dch: ignore