In Processable::Group, use new Processable subclasses; improve error handling.
This instantiates the five users of the mew Moo role Processable,
which properly abstracts the package type. The new role is completely
type-agnostic.

Also eliminates an errorhandler related to an 'eval' in Unpacker. The
eval there was removed. It trapped errors but also led to the obsuring
of errors. Lintian failed later on, but with the wrong error messages.
The last change may improve our understanding of why lindsay.d.o
sometimes bails out.

Also clarifies the boolean return type in Group->unpack for one code
path.

Gbp-Dch: ignore