Split Processable into five role users; Processable becomes a properly abstracted Moo role.
The five consumers of the new Moo role Processable deal with most type
related issues, so that much of the code base surrounding Processable
can become type-agnostic.

The notable exception is presently Processable::Group, which will
ensure going forward that, for source packages, all files are
symlinked. That may be revisited. It works great for now.

Makes the mechanism for guessing package names from file names more
explict via 'basename'.

That code path is difficult to test because it is rarely taken.  In
the test suite, only the 'fields-general-missing' family is confirmed
to require a guess. Those cases is furtermore special because there is
no underscore; the fields for version and architecture are missing.

Gbp-Dch: ignore