In Unpacker, eliminate an 'eval' that obscured messages to improve error handling.
Eliminates an 'eval' which trapped errors but sometimes led to the
obsuring of hideous errors. Lintian failed later on, but with the
wrong error messages!

This change may improve our understanding of why lindsay.d.o sometimes
bails out. It could also negatively affect the ability of automated
runs to recover.