In collection scripts, use new Collect::Dispatcher.
The collection scripts are not yet integrated into their respective
roles for access to collected data. Until that happens, they have to
instantiate their own Collect class objects, which they do with the
help of the new Dispatcher module.

This commit is part of a series to group data generation (in the
collection scripts) with data consumption (in the Collect class).  The
vehicles for that regrouping are Perl roles for each of the data
sources. The relative dependencies between those roles will describe
(and perhaps automatically load) the required prerequisites.

Gbp-Dch: ignore