Adjust Lintian::Path to new instantiation method for Collect::Source objects.
The source-related Collect object can be instatiated directly here.
There is no need to go through dispatcher since the processable type
is always 'source'.

This commit is part of a series to group data generation (in the
collection scripts) with data consumption (in the Collect class).  The
vehicles for that regrouping are Perl roles for each of the data
sources. The relative dependencies between those roles will describe
(and perhaps automatically load) the required prerequisites.

Gbp-Dch: ignore