Require libnamespace-clean-perl in both Depends and Build-Depends.
Moo::Role modifies the symbol table, but the behaviour is sensitive to
the order of imports in the consuming code. That caused a build error
for testing an unstable on Gitlab, which this commit series tries to
fix.

The build error did not appear on Debian stable before the offending
changes were committed to the group repo.

Stable and testing have the same version of Moo, so the change is
presumably in Perl, which is different, or in constant.pm.

I had followed a recommendation to place Moo::Role last, but that did
not clean the namespace. The POD tests then picked up the offending
names (all constants) from the symbol table.

Moo's behavior is documented here:

    https://metacpan.org/pod/Moo#CLEANING-UP-IMPORTS

This adds namespace::clean to Lintian's prerequisites in d/control.

A ticket was filed under RT#8015 to also update the packages installed
on lindsay.d.o