In consuming packages, place Moo::Role last among imports and clean up the namespace.
Newer versions of Debian (testing and unstable) trigger the POD tests
because the namespace is not clean.

The issue is described here:

    https://metacpan.org/pod/Moo#CLEANING-UP-IMPORTS

This commit cleans the namespace in modules that use Moo::Role.  If
that solves the build failures on Gitlab, a similar technique will be
applied to regular Moo imports, as well.

Gbp-Dch: ignore