Use new Dispatcher::create_info in Processable.
Replaces the old if-elsif structure to instantiate the right Collect
object. When the Collect::Dispatcher class goes away, the if-elsif
structure from there will be inserted here, and only here.

This commit is part of a series to group data generation (in the
collection scripts) with data consumption (in the Collect class).  The
vehicles for that regrouping are Perl roles for each of the data
sources. The relative dependencies between those roles will describe
(and perhaps automatically load) the required prerequisites.

Gbp-Dch: ignore