In Collect::Changes, use new feature roles to acquire collection capabilities.
Consumes the appropriate roles to provide access to all data needed to
analyze 'changes' files.

This commit is part of a series to group data generation (in the
collection scripts) with data consumption (in the Collect class).  The
vehicles for that regrouping are Perl roles for each of the data
sources. The relative dependencies between those roles will describe
(and perhaps automatically load) the required prerequisites.

Gbp-Dch: ignore