Turn Lintian::Collect into a Moo common role for all consumers of collected data.
This role provides all the basic and common data structures for
consumers of data generated by collection scripts.

This commit is part of a series to group data generation (in the
collection scripts) with data consumption (in the Collect class).  The
vehicles for that regrouping are Perl roles for each of the data
sources. The relative dependencies between those roles will describe
(and perhaps automatically load) the required prerequisites.

Gbp-Dch: ignore.