Turn Collect::Package into a Moo role to access package payloads; rename to Info::Package.
This role will be used by all consumers of package data, i.e. data for
processables with a payload, such as deb, udeb, and dsc. It provides
the data structures, the content and the accessors for the collected
data.

This commit is part of a series to group data generation (in the
collection scripts) with data consumption (in the Collect class).  The
vehicles for that regrouping are Perl roles for each of the data
sources. The relative dependencies between those roles will describe
(and perhaps automatically load) the required prerequisites.

Gbp-Dch: ignore