With Collect::Dispatcher, add a new module to automatically select the right type of Collect object.
This module serves as a central convenience while there are still
multiple instantiators of Collect class objects. It will become
obsolete after the collection scripts are properly instantiated for
each package processed.

This commit is part of a series to group data generation (in the
collection scripts) with data consumption (in the Collect class).  The
vehicles for that regrouping are Perl roles for each of the data
sources. The relative dependencies between those roles will describe
(and perhaps automatically load) the required prerequisites.

Gbp-Dch: ignore