Add new role Info::Checksums::Md5 to convey access to collected md5sums data.
This new role will be used by all consumers of md5sums data. It
provides the data structures, the content and the accessors for the
collected data.

This commit is part of a series to group data generation (in the
collection scripts) with data consumption (in the Collect class).  The
vehicles for that regrouping are Perl roles for each of the data
sources. The relative dependencies between those roles will describe
(and perhaps automatically load) the required prerequisites.

Gbp-Dch: ignore