In Collect::Dispatcher, use constructors for Lintian::Processable instead of Lintian::Collect.
All collected data for Processable objects is now held inside it. The
separate 'info' class Lintian::Collect ceased to exist. This
instantiates the objects of the new type in Collect::Dispatcher.

For clarity, the Collect::Dispatcher class could likewise move to
Processable::Dispatcher except that doing so now would require to edit
all the imports in collection scripts. With upcoming changes,
Collect::Dispatcher may disappear entirely, so the move was postponed
for now.

Gbp-Dch: ignore