Disable t/scripts/needs-info-missing.t; Needs-Info documentation logic no longer works in new class structure.
Only those function in Lintian::Processable that retrieve collected
information should require the Needs-Info line in POD, but the test
script requires it for all.

This topic should be revisited once all the functionality found its
final logical place. It's possible that dependency relationships
between the different collection items will in the future be
determined automatically and on-demand. That may render the explit
documentation obsolete.

Gbp-Dch: ignore