Turn the soon-to-be-obsolete modules in Lintian::Collect::* into Moo roles to be consumed by Processable.
For a limited time, these modules will provide access to data
facilities not yet offered by Lintian::Info. For that duration, they
will be Moo roles for use in Processable.

The code in these modules will be moved to the collection scripts,
which now create temporary databases instead of flat files.  (Required
to solve Bug#929729.) The databases are read in Lintian::Info, but the
data exchange will soon take place in memory---once the stubborn fork
in Unpacker is gone.

Gbp-Dch: ignore