In Lintian::Processable::*, consume appropriate Collect roles and set duplicate attributes.
Uses the appropriate Collect::* roles to ensure data transfer until
the migration to the data-centric Lintian::Info is complete.

Sets duplicate attributes that had different names in Collect and
Processable, such as name and pkg_name or type and pkg_type.

We could have used MooX::Aliases to avoid the assignments but it was
too cumbersome to update the prerequisites for linsday with DSA. The
situation is temporary until the code base is refactored for a stable
set of attribute names.

Gbp-Dch: ignore