Move Lintian::Collect->index to Collect::Binary to avoid competition with Collect::Source->index.
This method was in the wrong module. It relied on the order of imports
override one with the other. The problem became apparent when
converting to Moo roles.

The differences Collect->index and Source->index looked very small but
have to be worked out before one method can replace the other.

It is fixed here.

Gbp-Dch: ignore