In t/scripts/pod-coverage.t, drop reference to obsolete class Lintian::Collect.
This class was merged into Lintian::Processable and no longer exists.

This error did not appear on a desktop computer with a prior version
of Lintian installed because Perl will load the missing module from
the system path even though it was deleted from the development tree.

The POD coverage test hardcodes a whitelist of modules. A better way
would be to use Pod::Coverage::TrustPod. It will be implemented in the
near future.

This should restore building in Gitlab.

Gbp-Dch: ignore