Insert 'keys' before hashes for Perl to fix stretch-bpo. (Closes: #943910)
Older versions of Perl used a different hash signature when coercing
to a scalar. It returned a string describing the hash instead of the
number of elements. That caused a build failure in stretch-bpo. It is
addressed here.

The behavior of old Perls is described here:

    https://perldoc.perl.org/perl5260delta.html#scalar(%25hash)-return-signature-changed

In a nutshell, the following command produces different return values
in buster and stretch:

    stretch$ perl -e '%a = (1,2,3); print scalar %a, " -- ", scalar keys %a, "\n";'
    2/8 -- 2

    buster$ perl -e '%a = (1,2,3); print scalar %a, " -- ", scalar keys %a, "\n";'
    2 -- 2

This inserts 'keys' in the two known locations where it is missing.
For more information, please consult Bug#943910.

Thank you to Andreas Beckmann for reporting this issue and for taking
the time to identify, propose and confirm this fix.