In checks, clarify name of group directory; use Path::Tiny for path calculations.
The name of the directory 'basedir' was confusing because it was used
for the Lab, the Pool and for each group. This change clarifies when
the group directory is being used.

The group directory is the holding directory for all the analysis
related to a 'group' of files, which ordinarily derive from the same
source package and have the same version.

Also, a 'convenience function' for path calculations in Processable
was removed. This uses Path::Tiny for path calculations instead.

It may be possible to use string concatenation to avoid overhead and
module load times, except the author recently found a bug in Perl in
that circumstance:

    https://github.com/Perl/perl5/issues/17254

Path::Tiny is well-tested module that may come in handy for other file
or path related operations in checks in the future.

Gbp-Dch: ignore