Remove obsolete module Unpacker; was folded into Processable::Group.
The Unpacker was an attempt to parallelize the processing of Lintian
inputs, but it did so only when collecting information. A more
comprehensive approach would also multi-task the processing of groups.

Ultimately, the Unpacker was never able to acquire a greater scope
than a single group. It was only used in one instance in
Processable::Group but led to a needless shuffling of data and
execution between the two modules. In the end it was best to combine
the two.

There is a recalcitrant fork in the old Unpacker code (now part of
Processable::Group) that prevents the full use of IO::Async across
that process boundary. Once that has been resolved, the data exchange
between the collection scripts and the main program would no longer
need to involve files (or databases) but would take place in memory.

Until then, the collection of data is heavily constrained by disk use,
making the exclusive use of parallelization at that point
questionable. A better alternative may be to allow the processing of
Processable::Group objects to proceed in parallel, as well.

Gbp-Dch: ignore