In Processable, clarify name of group directory; remove obsolete convenience function.
The name of the directory 'basedir' was confusing because it was used
for the Lab, the Pool and for each group. This change clarifies when
the group directory is being used.

The group directory is the holding directory for all the analysis
related to a 'group' of files, which ordinarily derive from the same
source package and have the same version.

Also, a 'convenience function' for path calculations was removed. That
path calculation is better accomplished by the users of the group
directory.

Gbp-Dch: ignore