In Unpack::Task, clarify group directory; remove POD references to obsolete Lintian::Lab.
The name of the directory 'basedir' was confusing because it was used
for the Lab, the Pool and for each group. This change clarifies when
the group directory is being used.

The group directory is the holding directory for all the analysis
related to a 'group' of files, which ordinarily derive from the same
source package and have the same version.

Also removes large portions of obsolete POD documentation, including
several references to the module Lintian::Lab, which no longer exists.

Gbp-Dch: ignore