Adjust POD instructions to avoid test failures in preparation for relaxing the POD coverage test.
The old system used a whitelist and often caused test failures in
Gitlab when an obsolete class was listed but not found.

The error was impossible to detect when working on desktop computer
that had Lintian installed. Perl was able to resolve the missing
modules from the system path.

These edit were not made with great care or patience. Their sole
purpose was to avoid test errors. Future improvements are recommended.

Gbp-Dch: ignore