Test for POD coverage in all files; allow exceptions via Pod::Coverage::TrustPod.
The old system used a whitelist and often caused test failures in
Gitlab because an obsolete class was listed but not found.

The error was impossible to detect when working on desktop computer
that had Lintian installed. Perl was able to resolve the missing
modules from the system path.

This provides a more flexible system, in which exceptions to POD
coverage testing are noted in the affected files themselves.