Amend Lintian's manpage with '--hyperlinks' for text-based hyperlinks on terminals that support it.
A small number of terminals supports text-based hyperlinks. In Linux,
only the GNOME Terminal is supported. This commit offers a command
line option '--hyperlinks' that turns those hyperlinks on or off.

For the time being, only the tag name is linked to a page offering a
description and statistics on lintian.d.o.

The option is off on terminals that do not support it.  Otherwise, it
is on by default unless Lintian is invoked with '--color never'.

More information about the feature is available here:

    https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda

and at the bottom of this post:

    https://purpleidea.com/blog/2018/06/29/hyperlinks-in-gnome-terminal/

Gbp-Dch: ignore