In Processable::Source, consume role Info::Fields::Files.
The new role replaces the method 'files' that was originally obtained
from Collect::Source, which was organized by type of processable.

The Info module tree, on the other hand, is data-centric but knows
about processable types on the side.

The advantage of the new role organization is that code in the same
place relates to the same data source. It is no longer spread out
between different type-centric roles or subclasses.

Gbp-Dch: ignore