In Collect::Source, remove method 'files' which was split out as Info::Fields::Files.
This method is no longer needed. The processable can get it from
elswhere. The Info module tree is data-centric and knows about
processable types.

The advantage is that the modules there relate to a particular data
source and are no longer spread between different type-centric roles
or subclasses.

Removes the method.

Gbp-Dch: ignore