Drop minimum version requirement 8.30 for coreutils; process files containing newlines separately. (Closes #944258)
Commit b8e9d048 imposed a minimum version requirement for coreutils
(>= 8.30) but backporting coreutils was too much of a change for
stretch.  The requirement is being removed.

This reverts commit b8e9d04863aee57a1d9656a03bbd206525402d36.

Files whose names contain newlines are now processed separately with
md5sums. As of today, that step does not work with Jakub Wilk's
newline package from Bug#929729, but it will once the escaped file
names received from 'tar t' are properly reconstituted. For some
reason, escaped newlines are currently not translated back.

To be sure, the version requirement was imposed for good reason. The
solution for newlines that is presented here may not solve all the
questions raised by Bug#929729.

Initial experiments with exotic filenames show that their handling in
Lintian's collections remains defective, for example for invalid UTF-8
sequences. The files starting with 'bok' that are generated in
t/tags/checks/files/legacy-filenames/debian/rules are especially
problematic.

Some of the stability issues on lindsay.d.o may relate to file names
with international characters.

In a weighing of interests, however, it is much more important to
restore the benefits of Lintian's newest versions to users running
stretch than to accomodate unlikely filenames.

More information is available in the bug being closed. Bug#943910
has a report on building coreutils 8.30 for stretch.