Add new tag 'package-placeholder-in-symbols-file to detect the PACKAGE placeholder in d/symbols files.
This tag will be emitted when d/*.symbols files in source packages
contain the placeholder '#PACKAGE#'. The placeholder seems like a good
choice but is wrong.

Incorrect use of the Build-Depends-Package field in symbols files can
affect package dependencies. For that reason, the tag is emitted at
the the warning level.

For additional details, please have a look at Bug#944047.