In Lintian::Spelling, remove standard convenience tag emitter.
This tag emitter cannot use the new ->tag collector from which it was
called. The checks that need it have already set up their own tag
emitters. The routine here is no longer used. It is hereby removed.

This change is necessary to concentrate tag issuance in checks, which
makes the tags easier to post-process.

Gbp-Dch: ignore