In Lintian::Maintainer, do not issue tags via Tags::tag but return array of references to caller.
The tags that used to be issued here could not be routed through the
check-based post-processing and sorting scheme. This routine now
returns an array of tags that the caller, which is presumably a check,
can issue.

This change is necessary to concentrate tag issuance in checks, which
makes the tags easier to post-process.

Gbp-Dch: ignore