Move to next check on error, not to the next Processable.
The error is probably caused by a faulty check, so it is not necessary
to exclude the current processable altogether. A program exit may be
appropriate.

This is part of a series that dismantles global data structures for
tags issued, for overrides, and for associated statistics.  The
changes are intricate.

This commit passed the test suite on a desktop computer.

Gbp-Dch: ignore